ui: add bunch of comments to Control class and consolidate alot of 'Control' related enums into 'Control'
