+++ core/src/commonMain/kotlin/com/lehaine/littlekt/graph/node/node2d/ui/Container.kt
@@ -4,7 +4,6 @@ import com.lehaine.littlekt.graph.SceneGraph
 import com.lehaine.littlekt.graph.node.Node
 import com.lehaine.littlekt.graph.node.addTo
 import com.lehaine.littlekt.graph.node.annotation.SceneGraphDslMarker
-import com.lehaine.littlekt.graph.node.component.SizeFlag
 import com.lehaine.littlekt.graphics.Color
 import com.lehaine.littlekt.math.geom.Angle
 import com.lehaine.littlekt.util.Signal
